no point in making a header-only implementation here
