+++ http.h
@@ -13,27 +13,7 @@
 #ifndef DCUE_HTTP_H
 #define DCUE_HTTP_H
 
-#include "string_utility.h"
-
-#include <cstdint>
-#include <memory>
-#include <string>
-#include <vector>
-
-enum class HttpStatus { OK, NOT_FOUND, FORBIDDEN, INT_ERR, OTHER_FAIL };
-
-struct HttpHeader {
-  std::string name;
-  std::string value;
-};
-
-struct HttpResponse {
-  HttpStatus status;
-  std::vector<HttpHeader> headers;
-  std::vector<std::uint8_t> body;
-};
-
-#ifdef _WIN32
+#if 0
 #include "http_wininet.h"
 using HttpGet = HttpGetWinInet;
 #else
